refactor(dev): :recycle: move date validations
